真是遗憾,

+++ Question-Bank-and-Examination-System/src/main/java/justdj/top/dao/ExamMapper.java
@@ -95,5 +95,5 @@ public interface ExamMapper {
 			"where id = #{answerQuestionId}")
 	@Options(flushCache = true)
 	Integer updateAnswerQuestionScore(@Param("answerQuestionId")BigInteger answerQuestionId,
-	                                      @Param("score")Integer score);
+	                                      @Param("score")Short score);
 }
